requirements.txt
torch
transformers
Pillow
tqdm
timm
torchvision
